requirements.txt
# Airflow core
apache-airflow==2.0.2
pendulum==2.1.2

# Airflow provider packages (Udacity uses these for S3/Redshift)
apache-airflow-providers-amazon==1.3.0
apache-airflow-providers-postgres==1.0.1

# Compatible versions for Airflow 2.0.2
flask<2.0
werkzeug<2.1
connexion[swagger-ui]<2.8.0
packaging<21
jsonschema<4

# Extra utilities
boto3==1.17.27
psycopg2-binary==2.8.6
